refactor(catalog-cli): carve owners command family into catalog_cmds (nexus-kgyoz)
Seam 3 of the nexus-kgyoz god-object decomposition (epic nexus-whh61).
Carve the owner-management commands (owners, dedupe-owners) out of the
~6.9k-line commands/catalog.py into a new catalog_cmds/ package
(catalog_cmds/owners.py), the first split of that module. Commands are
defined as plain click commands plus a register(group) hook;
commands.catalog imports the submodule after the catalog group exists
and calls register(catalog), so 'nx catalog owners' / 'dedupe-owners'
resolve identically — names, options, and output unchanged.

Behaviour-preserving: owners_cmd reaches the shared _get_catalog helper
through the nexus.commands.catalog module object (not an import-time
binding) so the existing patch('nexus.commands.catalog._get_catalog')
test seam keeps working (seam-1 _cat_mod lesson). No test imports the
carved command functions by symbol; all invoke via the group.

Three contract pins in test_catalog_cli.py::TestSeam3OwnersCarve:
group-registration, callbacks-live-in-carved-module (anti-re-inline),
and a behavioural pin proving module-routed _get_catalog access.

Net -125 lines from the god module. Guard suites green (178 passed).
