+++ src/nexus/commands/catalog.py
@@ -227,6 +227,16 @@ def catalog() -> None:
     """
 
 
+# Command families carved out of this module live in catalog_cmds/ and attach
+# themselves to the group via their register() hook (nexus-kgyoz). Imported
+# here (after the group exists) so ``nx catalog owners`` / ``dedupe-owners``
+# resolve identically. catalog_cmds submodules reference this module's helpers
+# lazily, so this import stays acyclic.
+from nexus.commands.catalog_cmds import owners as _owners_cmds  # noqa: E402 — must follow the `catalog` group definition above
+
+_owners_cmds.register(catalog)
+
+
 @catalog.command("init")
 @click.option("--remote", default="", help="Optional git remote URL")
 def init_cmd(remote: str) -> None:
@@ -1600,33 +1610,6 @@ def link_audit_cmd(as_json: bool) -> None:
                 click.echo(f"  {o['from']} → {o['to']} ({o['type']})")
 
 
-@catalog.command("owners")
-@click.option("--json", "as_json", is_flag=True)
-def owners_cmd(as_json: bool) -> None:
-    """List registered owners."""
-    cat = _get_catalog()
-    owners = cat.list_owners()
-    if as_json:
-        data = [
-            {
-                "tumbler": o.get("tumbler_prefix"),
-                "name": o.get("name"),
-                "type": o.get("owner_type"),
-                "repo_hash": o.get("repo_hash"),
-                "description": o.get("description"),
-            }
-            for o in owners
-        ]
-        click.echo(json.dumps(data, indent=2))
-    else:
-        for o in owners:
-            click.echo(
-                f"{o.get('tumbler_prefix', ''):<8} "
-                f"{o.get('owner_type', ''):<10} "
-                f"{o.get('name', '')}"
-            )
-
-
 @catalog.command("sync")
 @click.option("--message", "-m", default="catalog update")
 def sync_cmd(message: str) -> None:
@@ -1637,104 +1620,6 @@ def sync_cmd(message: str) -> None:
     finally:
         cat.close()
     click.echo("Catalog synced.")
-
-
-@catalog.command("dedupe-owners")
-@click.option("--apply", is_flag=True, default=False,
-              help="Commit the plan. Default is dry-run.")
-@click.option("--json", "as_json", is_flag=True,
-              help="Emit the plan as JSON instead of a human summary.")
-def dedupe_owners_cmd(apply: bool, as_json: bool) -> None:
-    """Consolidate orphan owners (nexus-tmbh, part of nexus-b34f).
-
-    Classifies each curator owner as:
-
-    \b
-      • alias   — synthetic ``<repo>-<hash>`` names map to a canonical
-                  repo owner. Each doc is aliased via documents.alias_of
-                  to its canonical equivalent (matched by file_path).
-                  Rows stay so external references keep resolving.
-      • remove  — ``int-cce-*`` / ``int-prov-*`` / ``pdf-e2e-*`` test
-                  leakage predating RDR-060's autouse fixture. Documents,
-                  links, and the owner row are all deleted with JSONL
-                  tombstones.
-      • skip    — everything else (papers, knowledge, standalone-docs …).
-
-    Dry-run by default. Use ``--apply`` to commit, then ``nx catalog
-    sync`` to push the audit trail.
-    """
-    # Deep-maintenance: _dedupe.apply_plan mutates through the catalog's
-    # low-level event log / _db transactions, not the 22 daemon write ops
-    # (RDR-146). Use the full admin Catalog for both the plan read and the
-    # apply write.
-    from nexus.catalog.factory import (  # noqa: PLC0415 — deferred import; rare/branch-local path or circular-dep / startup-cost avoidance
-        CatalogAdminDaemonLiveError,
-        make_catalog_admin,
-    )
-    try:
-        cat = make_catalog_admin()
-    except CatalogAdminDaemonLiveError as exc:
-        raise click.ClickException(str(exc)) from exc
-    if cat is None:
-        raise click.ClickException(
-            "Catalog not initialized. Run 'nx catalog setup' first."
-        )
-    from nexus.catalog import dedupe as _dedupe  # noqa: PLC0415 — deferred import; rare/branch-local path or circular-dep / startup-cost avoidance
-
-    plan = _dedupe.plan_dedupe(cat)
-    summary = plan.summary()
-
-    if as_json:
-        payload = {
-            "dry_run": not apply,
-            "summary": summary,
-            "alias": [op.to_dict() for op in plan.alias],
-            "remove": [op.to_dict() for op in plan.remove],
-            "skip": [op.to_dict() for op in plan.skip],
-        }
-        if apply:
-            payload["applied"] = _dedupe.apply_plan(cat, plan)
-        click.echo(json.dumps(payload, indent=2))
-        return
-
-    label = "Would apply" if not apply else "Applying"
-    click.echo(f"{label} dedupe plan:")
-    click.echo(f"  alias:  {summary['alias']} owners, {summary['alias_docs']} docs")
-    click.echo(f"  remove: {summary['remove']} owners, {summary['remove_docs']} docs")
-    click.echo(f"  skip:   {summary['skip']} owners, {summary['skip_docs']} docs")
-
-    def _section(title: str, items: list, show_canonical: bool = False) -> None:
-        if not items:
-            return
-        click.echo(f"\n{title}:")
-        for op in items:
-            if show_canonical:
-                click.echo(
-                    f"  {op.orphan_prefix:<8} {op.orphan_name} "
-                    f"({op.doc_count} docs) → {op.canonical_prefix} {op.canonical_name}"
-                )
-            else:
-                click.echo(
-                    f"  {op.orphan_prefix:<8} {op.orphan_name} "
-                    f"({op.doc_count} docs)  — {op.reason}"
-                )
-
-    _section("Alias consolidation", plan.alias, show_canonical=True)
-    _section("Unconditional removal", plan.remove)
-    _section("Skipped (manual review)", plan.skip)
-
-    if not apply:
-        click.echo("\nDry-run only. Re-run with --apply to commit, "
-                   "then `nx catalog sync` to push the audit trail.")
-        return
-
-    totals = _dedupe.apply_plan(cat, plan)
-    click.echo("\nApplied:")
-    click.echo(f"  orphans aliased:  {totals['orphans_aliased']} "
-               f"({totals['aliased_docs']} docs, {totals['unmatched_docs']} unmatched)")
-    click.echo(f"  orphans removed:  {totals['orphans_removed']} "
-               f"({totals['removed_docs']} docs, {totals['removed_links']} links)")
-    click.echo("\nRun `nx catalog sync` to commit and push the audit trail.")
 
 
 @catalog.command("pull")
